backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.12.0
qdrant-client==1.7.0
boto3==1.29.7
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.25.2
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
watchtower==3.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
psycopg2-binary==2.9.9
slowapi==0.1.9
tenacity==8.2.3
redis>=5.0.0

